Continue the monetizable refactoring
